requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
attrs==25.3.0
azure-common==1.1.28
azure-core==1.30.1
azure-mgmt-authorization==4.0.0
azure-mgmt-cognitiveservices==13.5.0
azure-mgmt-core==1.4.0
azure-mgmt-resource==23.0.1
azure-mgmt-search==9.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
distro==1.9.0
faiss-cpu==1.10.0
fastapi==0.104.0
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
hyperframe==6.1.0
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
jiter==0.9.0
jpype1==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
konlpy==0.6.0
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.54
langchain-text-splitters==0.3.8
langsmith==0.3.32
llvmlite==0.44.0
lxml==5.3.2
MarkupSafe==3.0.2
marshmallow==3.26.1
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.4.3
mypy-extensions==1.0.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.5
openai==1.55.3
openai-whisper==20240930
orjson==3.10.16
packaging==24.2
portalocker==2.10.1
propcache==0.3.1
protobuf==5.29.4
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pydub==0.25.1
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
qdrant-client==1.7.3
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.27.0
sympy==1.13.1
tenacity==9.1.2
tiktoken==0.9.0
torch==2.6.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.23.2
yarl==1.20.0
zstandard==0.23.0
